DECA-101: linting files and updated copyright
Fixed all of the tabs to spaces issues (it seems that if you use the aptana plugin you need to set the tabs to spaces option twice). Also fixed other issues that jslint was complainging about.

Also made sure that all of the included files had the copyright with the correct information.