DECA-101: Integration of subcomponents with cameraMessage page.
Made a bunch of changes to get the cameraMessage page working with the subcompoents. All of the cameraMessage page unit tests are currently passing, but more could still be added. Also need unit tests for the subcomponents. However this is, for the most part, a working cameraMessage component. Note that css styling hasn't been added yet.