DECA-50: Starting work on the imageRotater.
This will be used as a subcomponent of the left-right calibration to rotate the images. There are unit tests included which are all passing. The styling still needs to be done, but I have added the css needed to perform the actual rotation.Also will need to implment proper error handling for the ajax call.