DECA-101: Fixing up how the initialisation of the supportedCameras subcomponent
The component subcomponent is passed in it's model from the parent. Since the parent uses the same model, it saves on an ajax call by using the parent's. Also set css styling to not display the template for the supportedCameras component when the progress message is initially running. The template markup for the supportedCameras subcomponent was also out of date, so this too was updated.The component subcomponent is passed in it's model from the parent. Since the parent uses the same model, it saves on an ajax call by using the parent's. Also set css styling to not display the template for the supportedCameras component when the progress message is initially running. The template markup for the supportedCameras subcomponent was also out of date, so this too was updated.