DECA-101, DECA-50: Minor modifications to the imageRotater component, and continued work on cameraSetup
Moved the rotation calculation function out of the imageRotater component and into decapodUtils file. This way the cameraSetup and imageRotater component could share the code.

Changed the imageRotater component to append a timestamp to the end of the captured image's url to prevent the browser from caching it.

Updated the unit tests for the imageRotater component.

Stabalized the cameraSetup component and now have all unit tests passing. The next step will be to complete the migration over to the calibration page.