DECA-132: Refactoring image rotation.
Remove the decapod.rotationInDegrees function from the decapodUtils.js file and placed it as a private function in imageRotater.

Now the cameraSetup component does not handle any rotation logic, this is all handled by the imageRotater.
This means that cameraSetup is expecting an unrotated image back from the server, and it will pass along the servers current rotation
information to the imageRotater component