DECA-183: Starting the Image Management page, added a mock book for local testing
Implemented the start of the image management page. It so far renders out the thumbnails and preview based on a model. The current model is read in from the local file system, where there is a mock-book for local testing. There are no events yet. Those should be coming in a future commit. The reordering of the thumbnails is currenlty commented out as there appears to be a bug with the infusion code that initializes it. Hope to have that sorted out shortly.