DECA-183: Continuing work on imageManagement page
Added in string templates to be used for the page labels and alt text, as well as for the Label of the Delete button.

Wired up events for broadcasting when a thumbnail or its delete button have been clicked. Currently activatino of the thumbnail will show it in the preview. Delete has not be attached to anything yet, as the server side implementation of it still needs to be worked out.