DECA-190: Patially working ROA implemenation for decapod 0.5a
Most of the basic functionality for the RESTful Resource Oriented Architecture fore Decapod 0.5a has been implemented. However the export is failing on genpdf (still debugging). Also there are no tests yet, and many TODOs (e.g. raising appropriate Exceptions). The export is triggered asynchronously though and the UI has been updated to import images to the new URL. There is also a lot of cleanup to remove old code that needs to happen still and lots of refactoring too.